requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
psycopg2==2.9.9
py==1.11.0
pyasn1==0.5.1
pydantic==2.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.3
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
bcrypt==4.0.1